backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.7.0
pymongo==4.10.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.18
passlib[bcrypt]==1.7.4
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
slowapi==0.1.9
cryptography==44.0.0
python-dotenv==1.0.1
